+++ src/pyfoma/_private/eliminate_flags.py
