+++ src/pyfoma/_private/transition.py
@@ -1,0 +1,6 @@
+class Transition:
+    __slots__ = ['targetstate', 'label', 'weight']
+    def __init__(self, targetstate, label, weight):
+        self.targetstate = targetstate
+        self.label = label
+        self.weight = weight
