+++ src/pyfoma/_private/util.py
@@ -1,0 +1,8 @@
+import subprocess
+
+def check_graphviz_installed():
+    try:
+        subprocess.run(["dot", "-V"], check=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+        return True
+    except (subprocess.CalledProcessError, FileNotFoundError):
+        return False
